Rework types in the loadbalancer
We used to rely on the invokers being indexed by number. That assumption is lifted and invokers now simply have names (sorting still relies on the numbering for now).

- Normalizing kafka topics to equal the name of the invoker
- Removing mutable datastructures
- Removing misleading Long (sizes are Int anyway)

+++ core/controller/src/main/scala/whisk/core/entitlement/ActivationThrottler.scala
@@ -50,7 +50,7 @@ class ActivationThrottler(consulServer: String, loadBalancer: LoadBalancer, conc
      * the number of concurrent invocations it has in the system
      */
     @volatile
-    private var userActivationCounter = Map.empty[String, Long]
+    private var userActivationCounter = Map.empty[String, Int]
 
     private val healthCheckInterval = 5.seconds
     private val consul = new ConsulClient(consulServer)
@@ -58,7 +58,7 @@ class ActivationThrottler(consulServer: String, loadBalancer: LoadBalancer, conc
     /**
      * Checks whether the operation should be allowed to proceed.
      */
-    def check(subject: Subject): Boolean = userActivationCounter.getOrElse(subject.asString, 0L) < concurrencyLimit
+    def check(subject: Subject): Boolean = userActivationCounter.getOrElse(subject.asString, 0) < concurrencyLimit
 
     /**
      * Checks whether the system is in a generally overloaded state.
