+++ tests/src/test/scala/whisk/core/controller/test/ControllerTestCommon.scala
@@ -189,7 +189,7 @@ class DegenerateLoadBalancerService(config: WhiskConfig)(implicit ec: ExecutionC
     // unit tests that need an activation via active ack/fast path should set this to value expected
     var whiskActivationStub: Option[(FiniteDuration, WhiskActivation)] = None
 
-    override def getActiveUserActivationCounts: Map[String, Long] = Map()
+    override def getActiveUserActivationCounts: Map[String, Int] = Map()
 
     override def publish(action: WhiskAction, msg: ActivationMessage, timeout: FiniteDuration)(implicit transid: TransactionId): Future[Future[WhiskActivation]] =
         Future.successful {
